compute reconstruction losses
One peculiarity- in most cases, the reconstruction loss when just using the training set's mean activation vector was slightly higher when using that to reconstruct the training set, compared with when using it to reconstruct the validation set. This recurs across most scenarios. I've double and triple checked the code and can't find any bugs that would cause this.